Added javadoc and Listing to be generic

+++ hw7/src/main/java/SaleContract.java
@@ -1,6 +1,15 @@
+/**
+ * A class that represents a sales contract
+ */
 public class SaleContract extends AbstractContract{
 
-  public SaleContract(Integer askingPrice, boolean negotiableOrNot) throws NegativeValueException {
+  /**
+   * Constructor for a sales contract
+   * @param askingPrice  the asking price, a non negative integer
+   * @param negotiableOrNot whether or not the price is negotiable, a boolean flag
+   * @throws NegativeValueException when asking price is negative
+   */
+  public SaleContract(Double askingPrice, boolean negotiableOrNot) throws NegativeValueException {
     super(askingPrice, negotiableOrNot);
   }
